python/calculators/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
ezdxf==0.17.2
gunicorn==20.1.0
loguru==0.6.0
numpy==1.22.4
numpy-stl==2.17.1
Pillow==9.1.1
psycopg2==2.8.6
pygame==2.0.0
pyparsing==3.0.9
python-utils==3.3.3
simple-draw==2.7.0
sqlparse==0.4.2
typing_extensions==4.2.0
whitenoise==6.2.0
